Finish EmbedWrapper
