+++ Aliucord/src/main/java/com/aliucord/wrappers/GuildWrapper.java
@@ -25,6 +25,7 @@ public class GuildWrapper {
         this.guild = guild;
     }
 
+    /** Returns the raw (obfuscated) {@link Guild} Object associated with this wrapper */
     public final Guild raw() {
         return guild;
     }
