+++ DiscordStubs/src/main/java/com/discord/api/message/embed/EmbedFooter.java
@@ -2,7 +2,7 @@ package com.discord.api.message.embed;
 
 /**
  * Obfuscated class with regularly changing method names.
- * Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedFooterWrapper} instead.
+ * Do not use this directly, use {@link com.aliucord.wrappers.embeds.FooterWrapper} instead.
  */
 @SuppressWarnings("unused")
 public final class EmbedFooter {
@@ -12,13 +12,13 @@ public final class EmbedFooter {
 
     /**
      * getProxyIconUrl
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedFooterWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.FooterWrapper} instead.
      */
     @Deprecated
     public final String a() { return proxyIconUrl; }
     /**
      * getText
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedFooterWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.FooterWrapper} instead.
      */
     @Deprecated
     public final String b() { return text; }
