+++ DiscordStubs/src/main/java/com/discord/api/message/embed/EmbedImage.java
@@ -2,7 +2,7 @@ package com.discord.api.message.embed;
 
 /**
  * Obfuscated class with regularly changing method names.
- * Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedImageWrapper} instead.
+ * Do not use this directly, use {@link com.aliucord.wrappers.embeds.ImageWrapper} instead.
  */
 @SuppressWarnings("unused")
 public final class EmbedImage {
@@ -13,25 +13,25 @@ public final class EmbedImage {
 
     /**
      * getHeight
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedImageWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ImageWrapper} instead.
      */
     @Deprecated
     public final Integer a() { return height; }
     /**
      * getProxyUrl
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedImageWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ImageWrapper} instead.
      */
     @Deprecated
     public final String b() { return proxyUrl; }
     /**
      * getUrl
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedImageWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ImageWrapper} instead.
      */
     @Deprecated
     public final String c() { return url; }
     /**
      * getWidth
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper.EmbedImageWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ImageWrapper} instead.
      */
     @Deprecated
     public final Integer d() { return width; }
