+++ DiscordStubs/src/main/java/com/discord/api/message/embed/EmbedProvider.java
@@ -1,12 +1,22 @@
 package com.discord.api.message.embed;
 
+/**
+ * Obfuscated class with regularly changing method names.
+ * Do not use this directly, use {@link com.aliucord.wrappers.embeds.ProviderWrapper} instead.
+ */
 @SuppressWarnings("unused")
 public final class EmbedProvider {
     private String name;
     private String url;
 
-    /** getName */
+    /**
+     * getName
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ProviderWrapper} instead.
+     */
     public final String a() { return name; }
-    /** getUrl */
+    /**
+     * getUrl
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.ProviderWrapper} instead.
+     */
     public final String b() { return url; }
 }
