+++ DiscordStubs/src/main/java/com/discord/api/message/embed/MessageEmbed.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 /**
  * Obfuscated class with regularly changing method names.
- * Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+ * Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
  */
 @SuppressWarnings("unused")
 public final class MessageEmbed {
@@ -26,79 +26,79 @@ public final class MessageEmbed {
 
     /**
      * getAuthor
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedAuthor a() { return author; }
     /**
      * getColor
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final Integer b() { return color; }
     /**
      * getDescription
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final String c() { return description; }
     /**
      * getFields
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final List<EmbedField> d() { return fields; }
     /**
      * getFooter
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedFooter e() { return footer; }
     /**
      * getImage
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedImage f() { return image; }
     /**
      * getProvider
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedProvider g() { return provider; }
     /**
      * getThumbnail
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedThumbnail h() { return thumbnail; }
     /**
      * getTimestamp
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final UtcDateTime i() { return timestamp; }
     /**
      * getTitle
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final String j() { return title; }
     /**
      * getType
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedType k() { return type; }
     /**
      * getUrl
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final String l() { return url; }
     /**
      * getVideo
-     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.MessageEmbedWrapper} instead.
+     * @deprecated Do not use this directly, use {@link com.aliucord.wrappers.embeds.MessageEmbedWrapper} instead.
      */
     @Deprecated
     public final EmbedVideo m() { return video; }
